Move node tool into utils
As it's used by both `selenium.py` and `services.py`, but they shouldn't
depend on one another.
